@@ -1,6 +1,6 @@
 package com.learning.core.exceptions;
 
-import javax.persistence.EntityNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 
 public class PasswordResetNotFound extends EntityNotFoundException {
 
