@@ -1,8 +1,8 @@
 package com.learning.web.dtos;
 
 
-import javax.validation.constraints.NotEmpty;
-import javax.validation.constraints.NotNull;
+import jakarta.validation.constraints.NotEmpty;
+import jakarta.validation.constraints.NotNull;
 import com.learning.web.interfaces.IConfirmPassword;
 import lombok.AllArgsConstructor;
 import lombok.Data;
