@@ -1,22 +1,22 @@
-//package com.learning.core.services;
-//import org.springframework.beans.factory.annotation.Autowired;
-//import org.springframework.stereotype.Service;
-//import br.com.treinaweb.ediaristas.core.exceptions.TokenNaBlackListException;
-//import br.com.treinaweb.ediaristas.core.models.TokenBlackList;
-//import br.com.treinaweb.ediaristas.core.repositories.TokenBlackListRepository;
-//
-//@Service
-//public class TokenBlackListService {
-//
+// package com.learning.core.services;
+// import org.springframework.beans.factory.annotation.Autowired;
+// import org.springframework.stereotype.Service;
+
+// import com.learning.core.exceptions.TokenNaBlackListException;
+// import com.learning.core.repository.TokenBlackListRepository;
+
+// @Service
+// public class TokenBlackListService {
+
 //    @Autowired
 //    private TokenBlackListRepository repository;
-//
+
 //    public void checkToken(String token) {
 //        if (repository.existsByToken(token)) {
 //            throw new TokenNaBlackListException();
 //        }
 //    }
-//
+
 //    public void colocarTokenNaBlackList(String token) {
 //        if (!repository.existsByToken(token)) {
 //            var tokenBlackList = new TokenBlackList();
@@ -24,5 +24,5 @@
 //            repository.save(tokenBlackList);
 //        }
 //    }
-//
-//}
+
+// }
