@@ -4,7 +4,7 @@ import com.learning.core.services.checkaddress.dtos.AddressResponse;
 import com.learning.core.services.checkaddress.exceptions.AddressServiceException;
 
 public interface AddressService {
-	
-	AddressResponse findAddressByCep(String cep) throws AddressServiceException;
+
+	AddressResponse findAddressByZipcode(String zipcode) throws AddressServiceException;
 
 }
